+++ source/wcx/WC6.cpp
@@ -231,7 +231,7 @@ namespace pksm
 
     u8 WC6::otFeeling(void) const { return data[0xE4]; }
 
-    u8 WC6::ev(Stat index) const { return data[0xE5 + u8(index)]; }
+    u16 WC6::ev(Stat index) const { return data[0xE5 + u8(index)]; }
 
     bool WC6::hasRibbon(Ribbon rib) const { return OFFSET_OF(rib).first != RIBBON_ABSENT; }
